csv version of train.json
This is the original uncleaned training data, however, it is not an exact copy of the original data because the table has been expanded. Specifically, the last column of train.json contains of lists of strings where each string is an ingredient. In this csv file, this last column has been expanded - replaced by many columns where each distinct ingredient was given its own column whose entries are 1 when "the ingredient was in the original list" and 0 when "the ingredient was not in the original list".